+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * New functions `mcmc_dots` and `mcmc_dots_by_chain` for dot plots of MCMC draws by @behramulukir (#402)
 * Default to `quantiles=100` for all dot plots by @behramulukir (#402)
+* Fix `mcmc_areas_ridges()` overlay breaking when `+ scale_y_discrete()` is added (#287)
 
 # bayesplot 1.15.0